+++ sshd-core/src/test/java/org/apache/sshd/common/config/TimeValueConfigTest.java
@@ -38,11 +38,11 @@ public class TimeValueConfigTest extends BaseTestSupport {
     @Test
     public void testDurationOf() {
         Object[] values = {
-            "600", Long.valueOf(TimeUnit.SECONDS.toMillis(600L)),
-            "10m", Long.valueOf(TimeUnit.MINUTES.toMillis(10L)),
-            "1h30m", Long.valueOf(TimeUnit.MINUTES.toMillis(90L)),
-            "2d", Long.valueOf(TimeUnit.DAYS.toMillis(2L)),
-            "3w", Long.valueOf(TimeUnit.DAYS.toMillis(3L * 7L))
+            "600", TimeUnit.SECONDS.toMillis(600L),
+            "10m", TimeUnit.MINUTES.toMillis(10L),
+            "1h30m", TimeUnit.MINUTES.toMillis(90L),
+            "2d", TimeUnit.DAYS.toMillis(2L),
+            "3w", TimeUnit.DAYS.toMillis(3L * 7L)
         };
         for (int index = 0; index < values.length; index += 2) {
             String s = (String) values[index];
